apps/bot/requirements.txt
# Core
flask>=3.0
python-dotenv>=1.0
aiohttp>=3.9

# Alpaca SDK
alpaca-py>=0.8

# Data / ML
pandas>=2.0
numpy>=1.24
torch>=2.2
optuna>=3.6

# Orchestration
langgraph>=0.2

# Scheduling
APScheduler>=3.10
